Added train to streamlit
